moving store.model file to the right folder and removing effects module
