+++ src/common/store.model.ts
