+++ src/pages/journal/journal.effects.ts
@@ -4,8 +4,8 @@ import { Actions, Effect, ofType } from '@ngrx/effects';
 import { switchMap, catchError, map } from 'rxjs/operators';
 import { of } from 'rxjs/observable/of';
 
-import * as journalActions from '../pages/journal/journal.actions';
-import { JournalProvider } from '../providers/journal/journal';
+import * as journalActions from './journal.actions';
+import { JournalProvider } from '../../providers/journal/journal';
 
 @Injectable()
 export class JournalEffects {
