+++ src/pages/journal/journal.module.ts
@@ -5,7 +5,7 @@ import { EffectsModule } from '@ngrx/effects';
 import { ComponentsModule } from '../../components/components.module';
 import { JournalPage } from './journal';
 import { journalReducer } from './journal.reducer';
-import { JournalEffects } from '../../effects/journal.effects';
+import { JournalEffects } from './journal.effects';
 
 @NgModule({
   declarations: [
